+++ src/CRON.go
@@ -5,6 +5,7 @@ import (
 	"io/ioutil"
 	"net/http"
 	"github.com/azukaar/cosmos-server/src/utils"
+	// "github.com/azukaar/cosmos-server/src/docker"
 	"os"
 	"path/filepath"
 	"encoding/json"
@@ -64,6 +65,8 @@ func checkVersion() {
 }
 
 func CRON() {
-	gocron.Every(1).Day().At("00:00").Do(checkVersion)
-	<-gocron.Start()
+	go func() {
+		gocron.Every(1).Day().At("00:00").Do(checkVersion)
+		<-gocron.Start()
+	}()
 }