+++ changelog.md
@@ -1,4 +1,17 @@
+## Version 0.1.16
+
+ - Fix search
+ - Fix bug where containers would lose their networks after being edited
+ - Self-heal secure network configuration
+ - Auto disconnect from orphan networks
+ - Prevent bootstrapping from creating orphan networks
+ - Monitor Docker and self-heal when docker daemon dies
+ - Recreate lost secure networks (ex. when resetting Cosmos)
+
+
+
 ## Version 0.1.15 
+
  - Ports is now freetype, in case container does not expose any
  - Container picker now tries to pick the best port as default
  - Hostname now default to container name